pyproject.toml
[project]
name = "missing-poster"
version = "0.1.2"
description = "Create funny Missing posters from pictures"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "pillow>=12.0.0",
]
[project.scripts]
missing-poster = "missing_poster.cli:main"

[tool.hatch.build]
include = ["src/yourpkg/**", "README.md", "LICENSE"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[project.urls]
Homepage = "https://github.com/moemchen13/missing_poster"
Repository = "https://github.com/moemchen13/missing_poster"
Issues = "https://github.com/moemchen13/missing_poster/issues"